+++ src/Domains/Intelligence/Agents/Laravel/Tools/Templates/GenerateTemplatePdfTool.php
@@ -9,8 +9,8 @@ use Kanvas\Intelligence\Agents\Attributes\AgentTool;
 use Kanvas\Intelligence\Agents\Laravel\Contracts\HasEntityContextInterface;
 use Kanvas\Intelligence\Agents\Laravel\Contracts\KanvasToolInterface;
 use Kanvas\Intelligence\Agents\Laravel\Traits\HandlesToolRequest;
-use Kanvas\Intelligence\Agents\Laravel\Traits\HasEntityContext;
 use Kanvas\Intelligence\Agents\Laravel\Traits\HasKanvasContext;
+use Kanvas\Intelligence\Agents\Traits\HasEntityContext;
 use Kanvas\Intelligence\Tools\Traits\Templates\ManagesTemplatesTrait;
 use Laravel\Ai\Tools\Request;
 use Override;
